Add new book to the API
