+++ package.json
@@ -14,6 +14,7 @@
     "react-redux": "^8.1.0",
     "react-router-dom": "^6.12.0",
     "react-scripts": "5.0.1",
+    "uuid": "^9.0.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
