+++ src/components/AddNewBook.js
@@ -1,10 +1,9 @@
 import { useDispatch, useSelector } from 'react-redux';
 import { useState } from 'react';
-import { addBook } from '../redux/books/booksSlice';
+import { v4 as uuidv4 } from 'uuid';
+import { postBook } from '../redux/books/booksSlice';
 
 const AddNewBook = () => {
-  const newId = useSelector((state) => state.books.books.length + 1);
-
   const [title, settitle] = useState('');
   const [author, setauthor] = useState('');
 
@@ -20,8 +19,8 @@ const AddNewBook = () => {
 
   const onSubmit = (ev) => {
     dispatch(
-      addBook({
-        item_id: `item${newId}`,
+      postBook({
+        item_id: `item${uuidv4()}`,
         title,
         author,
         category: 'Fiction',
